[#159]Rename from LinksToTheWork to LinksToWork
